+++ routers/index.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const userRouter = require('./userRouter.router');
-const menuRouter = require('./menuRouter.router');
+const categoryRouter = require('./categoryRouter.router');
 
-router.use('/main', menuRouter);
+router.use('/main', categoryRouter);
 router.use('/users', userRouter);
 
 module.exports = router;
